refactor: move extractBounceDoc into static method
